# scratch/extracted_replit/Hello-World/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Libyan E-Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teachers/{teacherId}:
    get:
      operationId: getTeacher
      tags: [users]
      summary: Get teacher profile
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherProfile"
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        bioAr:
          type: string
          nullable: true
        language:
          type: string
          enum: [ar, en]
        hasPasskey:
          type: boolean
        biometricsVerified:
          type: boolean
        phoneVerified:
          type: boolean
        emailVerified:
          type: boolean
        balance:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - role
        - language
        - createdAt
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        fullNameAr:
          type: string
        bio:
          type: string
        bioAr:
          type: string
        avatarUrl:
          type: string
        language:
          type: string
          enum: [ar, en]
...
    TeacherProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        fullNameAr:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        bioAr:
          type: string
          nullable: true
        courseCount:
          type: integer
        studentCount:
          type: integer
        rating:
          type: number
        expertise:
          type: string
          nullable: true
      required:
        - id
        - fullName
        - fullNameAr
        - courseCount
        - studentCount
        - rating